// api/generated/models/createUsuarioDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Unibac API
 * Documentación de la API
 * OpenAPI spec version: 1.0
 */
import type { CreateUsuarioDtoNivel } from "./createUsuarioDtoNivel";
import type { CreateUsuarioDtoTipo } from "./createUsuarioDtoTipo";

export interface CreateUsuarioDto {
  /** @minLength 3 */
  usuario: string;
  descripcion?: string;
  /** @minLength 6 */
  clave: string;
  activo?: boolean;
  nivel: CreateUsuarioDtoNivel;
  tipo: CreateUsuarioDtoTipo;
  correo?: string;
  celular?: string;
}