// api/generated/models/permisoEnUsuarioResponseDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Unibac API
 * Autenticación: POST /auth/login fija la cookie HttpOnly access_token. En Swagger UI, ejecuta login en este mismo origen y luego las rutas protegidas enviarán la cookie automáticamente. Desde el frontend usa fetch/axios con credentials: include. Registro público: POST /auth/register-public solo si PUBLIC_REGISTRATION_ENABLED=true.
 * OpenAPI spec version: 1.0
 */
import type { AccionResponseDto } from "./accionResponseDto";
import type { ModuloResponseDto } from "./moduloResponseDto";

export interface PermisoEnUsuarioResponseDto {
  id: number;
  usuarioId: number;
  moduloId: number;
  accionId: number;
  modulo: ModuloResponseDto;
  accion: AccionResponseDto;
}